+++ SmartDeviceLink/SDLSyncMsgVersion.h
@@ -23,7 +23,7 @@
  */
 - (instancetype)initWithDictionary:(NSMutableDictionary *)dict;
 
-- (instancetype)initWithMajorVersion:(NSInteger)majorVersion minorVersion:(NSInteger)minorVersion;
+- (instancetype)initWithMajorVersion:(UInt8)majorVersion minorVersion:(UInt8)minorVersion;
 
 /**
  * @abstract The major version indicates versions that is not-compatible to previous versions
